Process limited number of RapidPro contacts - The RapidPro sync task is scheduled to run after every 5 mins RapidPro has a request limit of 2500 per hour. In order to avoid reaching the limit, there are utmost 8 calls per 5 mins interval. These calls can have data that will be hitting the RapidPro endpoints in a loop.
To avoid reaching the limit we can default to posting 25 contacts.
For each interval session that will be 8 by 25 = 200. 5 minutes per hour
means RapidProListener will be called 12 times making a total of 2400 requests.

It is advised not to schedule RapidPro tasks below 5 minutes. 25 is optimal
number of request for each call.